Refactor object tracking implementation by integrating ObjectTracker class and removing redundant code
